+++ backend/app/db.py
@@ -1,0 +1,19 @@
+from sqlmodel import Field, Session, SQLModel, create_engine, select
+from fastapi import Depends
+from typing import Annotated
+
+sqlite_file_name = "meet_tracker.db"
+sqlite_url = f"sqlite:///{sqlite_file_name}"
+
+connect_args = {"check_same_thread": False}
+engine = create_engine(sqlite_url, connect_args=connect_args)
+
+
+def create_db_and_tables():
+    SQLModel.metadata.create_all(engine)
+
+def get_session():
+    with Session(engine) as session:
+        yield session
+
+SessionDep = Annotated[Session, Depends(get_session)]
